update function to control inline mode container events
