+++ elements/itemfilter/src/methods/container/index.js
@@ -1,5 +1,6 @@
 export { default as searchFilterMethod } from "./search-filter";
 export { default as resetFilterMethod } from "./reset-filter";
+export { default as initPopoverMethod } from "./init-popver";
 export {
   handleFormClickMethod,
   handleShowDropdownOnFocusMethod,
